Remove the lmgtfy special message
When doing just a lmgtfy link, there was a message added to the results.  This was so that the website would behave in a certain way.  It bugs me, though, that it's in here.  I want to keep this just for tranformations. Whatver is using the data can do the handling.